Use save button to save DM relays
This needed me to use document.getElementById instead of Svelte states,
but it works.
