+++ src/routes/(app)/settings/nostr/+page.svelte
@@ -1,4 +1,4 @@
-<script>
+<script lang="ts">
   import { untrack } from "svelte";
   import { browser } from "$app/environment";
   import { getNsec } from "$lib/nostr";
@@ -17,9 +17,6 @@
   let npub = $state(untrack(() => user.npub));
   let extensionAvailable = $derived(browser && window.nostr);
   let { locale } = $derived(user);
-
-  import { SimplePool } from 'nostr-tools/pool';
-  import { finalizeEvent } from 'nostr-tools/pure';
 
   let newNsec = $state(),
     nsec = $state(),
@@ -44,28 +41,6 @@
     await tick();
     $save.click();
   };
-
-  const pool = new SimplePool();
-  import { PUBLIC_DM_RELAYS } from '$env/static/public';
-  const DM_RELAYS_LIST = PUBLIC_DM_RELAYS.split(',');
-  let dmRelaysText = $state('');
-  let updateRelays = async () => {
-    const newRelays = dmRelaysText.split('\n');
-    const event = {
-      kind: 10050,
-      created_at: Math.floor(Date.now() / 1000),
-      tags: newRelays.map(r => ["relay", r]),
-      content: ""
-    };
-    let signed;
-    if (await window.nostr.getPublicKey() === user.pubkey) {
-      signed = await window.nostr.signEvent(event);
-    } else {
-      const sk = await getPrivateKey(user);
-      signed = finalizeEvent(event, sk);
-    }
-    await Promise.any(pool.publish(DM_RELAYS_LIST, signed));
-  }
 </script>
 
 <input type="hidden" name="challenge" value={challenge} />
@@ -257,10 +232,11 @@
   </div>
 {/if}
 
-<p class="font-bold">
+<span class="font-bold">
     Preferred DM Relays
+</span>
+<p class="text-secondary mb-1">
+    Direct Messages meant for you will be published to these relays.
 </p>
-Direct Messages meant for you will be published to these relays.
 
-<textarea name="dmRelays" bind:value={dmRelaysText} rows={3}></textarea>
-<button onclick={updateRelays} class="btn">Update Preferred DM Relays</button>
+<textarea id="dmRelays" name="dmRelays" rows={3}></textarea>
